move action management to a separate ActionService
