+++ src/app/app.module.ts
@@ -78,13 +78,11 @@ import { MaterialModule } from './material.module';
 import { ExperimentalDirective } from './directives/experimental.directive';
 import { ContentApiService } from './services/content-api.service';
 import { ExtensionsModule } from './extensions.module';
-import { ExtensionService } from './extensions/extension.service';
-import { CoreExtensionsModule } from './extensions/core.extensions';
+import { CoreExtensionsModule } from './extensions/core.extensions.module';
 import { SearchResultsRowComponent } from './components/search/search-results-row/search-results-row.component';
 import { NodePermissionsDialogComponent } from './dialogs/node-permissions/node-permissions.dialog';
 import { NodePermissionsDirective } from './common/directives/node-permissions.directive';
 import { PermissionsManagerComponent } from './components/permission-manager/permissions-manager.component';
-import { RuleService } from './extensions/rules/rule.service';
 
 @NgModule({
     imports: [
@@ -159,9 +157,7 @@ import { RuleService } from './extensions/rules/rule.service';
         NodePermissionService,
         ProfileResolver,
         ExperimentalGuard,
-        ContentApiService,
-        ExtensionService,
-        RuleService
+        ContentApiService
     ],
     entryComponents: [
         LibraryDialogComponent,
