+++ src/app/extensions/actions/action.service.ts
@@ -1,0 +1,76 @@
+/*!
+ * @license
+ * Alfresco Example Content Application
+ *
+ * Copyright (C) 2005 - 2018 Alfresco Software Limited
+ *
+ * This file is part of the Alfresco Example Content Application.
+ * If the software was purchased under a paid Alfresco license, the terms of
+ * the paid license agreement will prevail.  Otherwise, the software is
+ * provided under the following open source license terms:
+ *
+ * The Alfresco Example Content Application is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * The Alfresco Example Content Application is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { Injectable } from '@angular/core';
+import { AppConfigService } from '@alfresco/adf-core';
+import { Store } from '@ngrx/store';
+import { AppStore } from '../../store/states';
+import { ActionRef } from './action-ref';
+
+@Injectable()
+export class ActionService {
+    actions: Array<ActionRef> = [];
+
+    constructor(
+        private config: AppConfigService,
+        private store: Store<AppStore>
+    ) {}
+
+    init() {
+        this.actions = this.config.get<Array<ActionRef>>(
+            'extensions.core.actions',
+            []
+        );
+    }
+
+    getActionById(id: string): ActionRef {
+        return this.actions.find(action => action.id === id);
+    }
+
+    runActionById(id: string, context?: any) {
+        const action = this.getActionById(id);
+        if (action) {
+            const { type, payload } = action;
+            const expression = this.runExpression(payload, context);
+
+            this.store.dispatch({ type, payload: expression });
+        }
+    }
+
+    runExpression(value: string, context?: any) {
+        const pattern = new RegExp(/\$\((.*\)?)\)/g);
+        const matches = pattern.exec(value);
+
+        if (matches && matches.length > 1) {
+            const expression = matches[1];
+            const fn = new Function('context', `return ${expression}`);
+            const result = fn(context);
+
+            return result;
+        }
+
+        return value;
+    }
+}
