+++ src/app/extensions/extension.config.ts
@@ -24,12 +24,12 @@
  */
 
 import { RouteExtension } from './route.extension';
-import { ActionExtension } from './action.extension';
 import { RuleRef } from './rules/rule-ref';
+import { ActionRef } from './actions/action-ref';
 
 export interface ExtensionConfig {
     rules?: Array<RuleRef>;
     routes?: Array<RouteExtension>;
-    actions?: Array<ActionExtension>;
+    actions?: Array<ActionRef>;
     features?: { [key: string]: any };
 }
