+++ src/app/extensions/extension.service.ts
@@ -25,24 +25,22 @@
 
 import { Injectable, Type } from '@angular/core';
 import { RouteExtension } from './route.extension';
-import { ActionExtension } from './action.extension';
 import { AppConfigService } from '@alfresco/adf-core';
 import {
     ContentActionExtension,
     ContentActionType
 } from './content-action.extension';
 import { OpenWithExtension } from './open-with.extension';
-import { AppStore } from '../store/states';
-import { Store } from '@ngrx/store';
 import { NavigationExtension } from './navigation.extension';
 import { Route } from '@angular/router';
 import { Node } from 'alfresco-js-api';
 import { RuleService } from './rules/rule.service';
+import { ActionService } from './actions/action.service';
+import { ActionRef } from './actions/action-ref';
 
 @Injectable()
 export class ExtensionService {
     routes: Array<RouteExtension> = [];
-    actions: Array<ActionExtension> = [];
 
     contentActions: Array<ContentActionExtension> = [];
     openWithActions: Array<OpenWithExtension> = [];
@@ -53,8 +51,8 @@ export class ExtensionService {
 
     constructor(
         private config: AppConfigService,
-        private store: Store<AppStore>,
-        private ruleService: RuleService
+        private ruleService: RuleService,
+        private actionService: ActionService
     ) {}
 
     // initialise extension service
@@ -62,11 +60,6 @@ export class ExtensionService {
     init() {
         this.routes = this.config.get<Array<RouteExtension>>(
             'extensions.core.routes',
-            []
-        );
-
-        this.actions = this.config.get<Array<ActionExtension>>(
-            'extensions.core.actions',
             []
         );
 
@@ -93,24 +86,19 @@ export class ExtensionService {
             .sort(this.sortByOrder);
 
         this.ruleService.init();
+        this.actionService.init();
     }
 
     getRouteById(id: string): RouteExtension {
         return this.routes.find(route => route.id === id);
     }
 
-    getActionById(id: string): ActionExtension {
-        return this.actions.find(action => action.id === id);
+    getActionById(id: string): ActionRef {
+        return this.actionService.getActionById(id);
     }
 
     runActionById(id: string, context?: any) {
-        const action = this.getActionById(id);
-        if (action) {
-            const { type, payload } = action;
-            const expression = this.runExpression(payload, context);
-
-            this.store.dispatch({ type, payload: expression });
-        }
+        this.actionService.runActionById(id, context);
     }
 
     getNavigationGroups(): Array<NavigationExtension[]> {
@@ -148,21 +136,6 @@ export class ExtensionService {
 
     getComponentById(id: string): Type<{}> {
         return this.components[id];
-    }
-
-    runExpression(value: string, context?: any) {
-        const pattern = new RegExp(/\$\((.*\)?)\)/g);
-        const matches = pattern.exec(value);
-
-        if (matches && matches.length > 1) {
-            const expression = matches[1];
-            const fn = new Function('context', `return ${expression}`);
-            const result = fn(context);
-
-            return result;
-        }
-
-        return value;
     }
 
     getApplicationRoutes(): Array<Route> {
